# lib/api-zod/src/generated/types/arcoAgenteEnvio.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Sistema de Romaneios
 * OpenAPI spec version: 0.1.0
 */

export interface ArcoAgenteEnvio {
  id: number;
  operation: string;
  romaneioLabel: string;
  romaneioData: string;
  codigos: string[];
}