# lib/api-zod/src/generated/types/arcoEnvio.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Sistema de Romaneios
 * OpenAPI spec version: 0.1.0
 */
import type { ArcoEnvioRomaneioEscopoTipo } from "./arcoEnvioRomaneioEscopoTipo";
import type { ArcoEnvioStatus } from "./arcoEnvioStatus";

export interface ArcoEnvio {
  id: number;
  operation: string;
  romaneioData: string;
  romaneioEscopoTipo: ArcoEnvioRomaneioEscopoTipo;
  romaneioEscopoValor: string;
  romaneioLabel: string;
  status: ArcoEnvioStatus;
  /** @nullable */
  criadoPor: string | null;
  criadoEm: string;
  /** @nullable */
  iniciadoEm: string | null;
  /** @nullable */
  finalizadoEm: string | null;
  /** @nullable */
  ultimoHeartbeat: string | null;
  /** @nullable */
  erroGeral: string | null;
  /** @nullable */
  envioOrigemId: number | null;
}