# lib/api-zod/src/generated/types/arcoEnvioComItens.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Sistema de Romaneios
 * OpenAPI spec version: 0.1.0
 */
import type { ArcoEnvio } from "./arcoEnvio";
import type { ArcoEnvioItem } from "./arcoEnvioItem";

export type ArcoEnvioComItens = ArcoEnvio & {
  itens: ArcoEnvioItem[];
};