# lib/api-zod/src/generated/types/arcoEnvioItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Sistema de Romaneios
 * OpenAPI spec version: 0.1.0
 */
import type { ArcoEnvioItemStatus } from "./arcoEnvioItemStatus";

export interface ArcoEnvioItem {
  id: number;
  envioId: number;
  codigo: string;
  status: ArcoEnvioItemStatus;
  /** @nullable */
  erro: string | null;
}